server refactoring 0.1
